third_party/blink/web_tests/fast/events/event-attribute-expected.txt

Test that setting event handlers with attribute works.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS /*img*/ typeof (element["onclick"]) is "function"
PASS /*img*/ typeof (element["oncontextmenu"]) is "function"
PASS /*img*/ typeof (element["ondblclick"]) is "function"
PASS /*img*/ typeof (element["onmousedown"]) is "function"
PASS /*img*/ typeof (element["onmousemove"]) is "function"
PASS /*img*/ typeof (element["onmouseout"]) is "function"
PASS /*img*/ typeof (element["onmouseover"]) is "function"
PASS /*img*/ typeof (element["onmouseup"]) is "function"
PASS /*img*/ typeof (element["onmousewheel"]) is "function"
PASS /*img*/ typeof (element["onfocus"]) is "function"
PASS /*img*/ typeof (element["onblur"]) is "function"
PASS /*img*/ typeof (element["onkeydown"]) is "function"
PASS /*img*/ typeof (element["onkeypress"]) is "function"
PASS /*img*/ typeof (element["onkeyup"]) is "function"
PASS /*img*/ typeof (element["onscroll"]) is "function"
PASS /*img*/ typeof (element["onbeforecut"]) is "function"
PASS /*img*/ typeof (element["oncut"]) is "function"
PASS /*img*/ typeof (element["onbeforecopy"]) is "function"
PASS /*img*/ typeof (element["oncopy"]) is "function"
PASS /*img*/ typeof (element["onbeforepaste"]) is "function"
PASS /*img*/ typeof (element["onpaste"]) is "function"
PASS /*img*/ typeof (element["ondragenter"]) is "function"
PASS /*img*/ typeof (element["ondragover"]) is "function"
PASS /*img*/ typeof (element["ondragleave"]) is "function"
PASS /*img*/ typeof (element["ondrop"]) is "function"
PASS /*img*/ typeof (element["ondragstart"]) is "function"
PASS /*img*/ typeof (element["ondrag"]) is "function"
PASS /*img*/ typeof (element["ondragend"]) is "function"
PASS /*img*/ typeof (element["onselectstart"]) is "function"
PASS /*img*/ typeof (element["onsubmit"]) is "function"
PASS /*img*/ typeof (element["onerror"]) is "function"
PASS /*img*/ typeof (element["oninput"]) is "function"
PASS /*img*/ typeof (element["oninvalid"]) is "function"
PASS /*img*/ typeof (element["ontouchstart"]) is "function"
PASS /*img*/ typeof (element["ontouchmove"]) is "function"
PASS /*img*/ typeof (element["ontouchend"]) is "function"
PASS /*img*/ typeof (element["ontouchcancel"]) is "function"
PASS /*img*/ typeof (element["onwebkitfullscreenchange"]) is "function"
PASS /*img*/ typeof (element["onwebkitfullscreenerror"]) is "function"
PASS /*img*/ typeof (element["onabort"]) is "function"
PASS /*img*/ typeof (element["onchange"]) is "function"
PASS /*img*/ typeof (element["onreset"]) is "function"
PASS /*img*/ typeof (element["onselect"]) is "function"
PASS /*img*/ typeof (element["onload"]) is "function"
PASS /*script*/ typeof (element["onclick"]) is "function"
PASS /*script*/ typeof (element["oncontextmenu"]) is "function"
PASS /*script*/ typeof (element["ondblclick"]) is "function"
PASS /*script*/ typeof (element["onmousedown"]) is "function"
PASS /*script*/ typeof (element["onmousemove"]) is "function"
PASS /*script*/ typeof (element["onmouseout"]) is "function"
PASS /*script*/ typeof (element["onmouseover"]) is "function"
PASS /*script*/ typeof (element["onmouseup"]) is "function"
PASS /*script*/ typeof (element["onmousewheel"]) is "function"
PASS /*script*/ typeof (element["onfocus"]) is "function"
PASS /*script*/ typeof (element["onblur"]) is "function"
PASS /*script*/ typeof (element["onkeydown"]) is "function"
PASS /*script*/ typeof (element["onkeypress"]) is "function"
PASS /*script*/ typeof (element["onkeyup"]) is "function"
PASS /*script*/ typeof (element["onscroll"]) is "function"
PASS /*script*/ typeof (element["onbeforecut"]) is "function"
PASS /*script*/ typeof (element["oncut"]) is "function"
PASS /*script*/ typeof (element["onbeforecopy"]) is "function"
PASS /*script*/ typeof (element["oncopy"]) is "function"
PASS /*script*/ typeof (element["onbeforepaste"]) is "function"
PASS /*script*/ typeof (element["onpaste"]) is "function"
PASS /*script*/ typeof (element["ondragenter"]) is "function"
PASS /*script*/ typeof (element["ondragover"]) is "function"
PASS /*script*/ typeof (element["ondragleave"]) is "function"
PASS /*script*/ typeof (element["ondrop"]) is "function"
PASS /*script*/ typeof (element["ondragstart"]) is "function"
PASS /*script*/ typeof (element["ondrag"]) is "function"
PASS /*script*/ typeof (element["ondragend"]) is "function"
PASS /*script*/ typeof (element["onselectstart"]) is "function"
PASS /*script*/ typeof (element["onsubmit"]) is "function"
PASS /*script*/ typeof (element["onerror"]) is "function"
PASS /*script*/ typeof (element["oninput"]) is "function"
PASS /*script*/ typeof (element["oninvalid"]) is "function"
PASS /*script*/ typeof (element["ontouchstart"]) is "function"
PASS /*script*/ typeof (element["ontouchmove"]) is "function"
PASS /*script*/ typeof (element["ontouchend"]) is "function"
PASS /*script*/ typeof (element["ontouchcancel"]) is "function"
PASS /*script*/ typeof (element["onwebkitfullscreenchange"]) is "function"
PASS /*script*/ typeof (element["onwebkitfullscreenerror"]) is "function"
PASS /*script*/ typeof (element["onabort"]) is "function"
PASS /*script*/ typeof (element["onchange"]) is "function"
PASS /*script*/ typeof (element["onreset"]) is "function"
PASS /*script*/ typeof (element["onselect"]) is "function"
PASS /*script*/ typeof (element["onload"]) is "function"
PASS /*video*/ typeof (element["onclick"]) is "function"
PASS /*video*/ typeof (element["oncontextmenu"]) is "function"
PASS /*video*/ typeof (element["ondblclick"]) is "function"
PASS /*video*/ typeof (element["onmousedown"]) is "function"
PASS /*video*/ typeof (element["onmousemove"]) is "function"
PASS /*video*/ typeof (element["onmouseout"]) is "function"
PASS /*video*/ typeof (element["onmouseover"]) is "function"
PASS /*video*/ typeof (element["onmouseup"]) is "function"
PASS /*video*/ typeof (element["onmousewheel"]) is "function"
PASS /*video*/ typeof (element["onfocus"]) is "function"
PASS /*video*/ typeof (element["onblur"]) is "function"
PASS /*video*/ typeof (element["onkeydown"]) is "function"
PASS /*video*/ typeof (element["onkeypress"]) is "function"
PASS /*video*/ typeof (element["onkeyup"]) is "function"
PASS /*video*/ typeof (element["onscroll"]) is "function"
PASS /*video*/ typeof (element["onbeforecut"]) is "function"
PASS /*video*/ typeof (element["oncut"]) is "function"
PASS /*video*/ typeof (element["onbeforecopy"]) is "function"
PASS /*video*/ typeof (element["oncopy"]) is "function"
PASS /*video*/ typeof (element["onbeforepaste"]) is "function"
PASS /*video*/ typeof (element["onpaste"]) is "function"
PASS /*video*/ typeof (element["ondragenter"]) is "function"
PASS /*video*/ typeof (element["ondragover"]) is "function"
PASS /*video*/ typeof (element["ondragleave"]) is "function"
PASS /*video*/ typeof (element["ondrop"]) is "function"
PASS /*video*/ typeof (element["ondragstart"]) is "function"
PASS /*video*/ typeof (element["ondrag"]) is "function"
PASS /*video*/ typeof (element["ondragend"]) is "function"
PASS /*video*/ typeof (element["onselectstart"]) is "function"
PASS /*video*/ typeof (element["onsubmit"]) is "function"
PASS /*video*/ typeof (element["onerror"]) is "function"
PASS /*video*/ typeof (element["oninput"]) is "function"
PASS /*video*/ typeof (element["oninvalid"]) is "function"
PASS /*video*/ typeof (element["ontouchstart"]) is "function"
PASS /*video*/ typeof (element["ontouchmove"]) is "function"
PASS /*video*/ typeof (element["ontouchend"]) is "function"
PASS /*video*/ typeof (element["ontouchcancel"]) is "function"
PASS /*video*/ typeof (element["onwebkitfullscreenchange"]) is "function"
PASS /*video*/ typeof (element["onwebkitfullscreenerror"]) is "function"
PASS /*video*/ typeof (element["onabort"]) is "function"
PASS /*video*/ typeof (element["onchange"]) is "function"
PASS /*video*/ typeof (element["onreset"]) is "function"
PASS /*video*/ typeof (element["onselect"]) is "function"
PASS /*video*/ typeof (element["onload"]) is "function"
PASS /*div*/ typeof (element["onclick"]) is "function"
PASS /*div*/ typeof (element["oncontextmenu"]) is "function"
PASS /*div*/ typeof (element["ondblclick"]) is "function"
PASS /*div*/ typeof (element["onmousedown"]) is "function"
PASS /*div*/ typeof (element["onmousemove"]) is "function"
PASS /*div*/ typeof (element["onmouseout"]) is "function"
PASS /*div*/ typeof (element["onmouseover"]) is "function"
PASS /*div*/ typeof (element["onmouseup"]) is "function"
PASS /*div*/ typeof (element["onmousewheel"]) is "function"
PASS /*div*/ typeof (element["onfocus"]) is "function"
PASS /*div*/ typeof (element["onblur"]) is "function"
PASS /*div*/ typeof (element["onkeydown"]) is "function"
PASS /*div*/ typeof (element["onkeypress"]) is "function"
PASS /*div*/ typeof (element["onkeyup"]) is "function"
PASS /*div*/ typeof (element["onscroll"]) is "function"
PASS /*div*/ typeof (element["onbeforecut"]) is "function"
PASS /*div*/ typeof (element["oncut"]) is "function"
PASS /*div*/ typeof (element["onbeforecopy"]) is "function"
PASS /*div*/ typeof (element["oncopy"]) is "function"
PASS /*div*/ typeof (element["onbeforepaste"]) is "function"
PASS /*div*/ typeof (element["onpaste"]) is "function"
PASS /*div*/ typeof (element["ondragenter"]) is "function"
PASS /*div*/ typeof (element["ondragover"]) is "function"
PASS /*div*/ typeof (element["ondragleave"]) is "function"
PASS /*div*/ typeof (element["ondrop"]) is "function"
PASS /*div*/ typeof (element["ondragstart"]) is "function"
PASS /*div*/ typeof (element["ondrag"]) is "function"
PASS /*div*/ typeof (element["ondragend"]) is "function"
PASS /*div*/ typeof (element["onselectstart"]) is "function"
PASS /*div*/ typeof (element["onsubmit"]) is "function"
PASS /*div*/ typeof (element["onerror"]) is "function"
PASS /*div*/ typeof (element["oninput"]) is "function"
PASS /*div*/ typeof (element["oninvalid"]) is "function"
PASS /*div*/ typeof (element["ontouchstart"]) is "function"
PASS /*div*/ typeof (element["ontouchmove"]) is "function"
PASS /*div*/ typeof (element["ontouchend"]) is "function"
PASS /*div*/ typeof (element["ontouchcancel"]) is "function"
PASS /*div*/ typeof (element["onwebkitfullscreenchange"]) is "function"
PASS /*div*/ typeof (element["onwebkitfullscreenerror"]) is "function"
PASS /*div*/ typeof (element["onabort"]) is "function"
PASS /*div*/ typeof (element["onchange"]) is "function"
PASS /*div*/ typeof (element["onreset"]) is "function"
PASS /*div*/ typeof (element["onselect"]) is "function"
PASS /*div*/ typeof (element["onload"]) is "function"
PASS /*form*/ typeof (element["onclick"]) is "function"
PASS /*form*/ typeof (element["oncontextmenu"]) is "function"
PASS /*form*/ typeof (element["ondblclick"]) is "function"
PASS /*form*/ typeof (element["onmousedown"]) is "function"
PASS /*form*/ typeof (element["onmousemove"]) is "function"
PASS /*form*/ typeof (element["onmouseout"]) is "function"
PASS /*form*/ typeof (element["onmouseover"]) is "function"
PASS /*form*/ typeof (element["onmouseup"]) is "function"
PASS /*form*/ typeof (element["onmousewheel"]) is "function"
PASS /*form*/ typeof (element["onfocus"]) is "function"
PASS /*form*/ typeof (element["onblur"]) is "function"
PASS /*form*/ typeof (element["onkeydown"]) is "function"
PASS /*form*/ typeof (element["onkeypress"]) is "function"
PASS /*form*/ typeof (element["onkeyup"]) is "function"
PASS /*form*/ typeof (element["onscroll"]) is "function"
PASS /*form*/ typeof (element["onbeforecut"]) is "function"
PASS /*form*/ typeof (element["oncut"]) is "function"
PASS /*form*/ typeof (element["onbeforecopy"]) is "function"
PASS /*form*/ typeof (element["oncopy"]) is "function"
PASS /*form*/ typeof (element["onbeforepaste"]) is "function"
PASS /*form*/ typeof (element["onpaste"]) is "function"
PASS /*form*/ typeof (element["ondragenter"]) is "function"
PASS /*form*/ typeof (element["ondragover"]) is "function"
PASS /*form*/ typeof (element["ondragleave"]) is "function"
PASS /*form*/ typeof (element["ondrop"]) is "function"
PASS /*form*/ typeof (element["ondragstart"]) is "function"
PASS /*form*/ typeof (element["ondrag"]) is "function"
PASS /*form*/ typeof (element["ondragend"]) is "function"
PASS /*form*/ typeof (element["onselectstart"]) is "function"
PASS /*form*/ typeof (element["onsubmit"]) is "function"
PASS /*form*/ typeof (element["onerror"]) is "function"
PASS /*form*/ typeof (element["oninput"]) is "function"
PASS /*form*/ typeof (element["oninvalid"]) is "function"
PASS /*form*/ typeof (element["ontouchstart"]) is "function"
PASS /*form*/ typeof (element["ontouchmove"]) is "function"
PASS /*form*/ typeof (element["ontouchend"]) is "function"
PASS /*form*/ typeof (element["ontouchcancel"]) is "function"
PASS /*form*/ typeof (element["onwebkitfullscreenchange"]) is "function"
PASS /*form*/ typeof (element["onwebkitfullscreenerror"]) is "function"
PASS /*form*/ typeof (element["onabort"]) is "function"
PASS /*form*/ typeof (element["onchange"]) is "function"
PASS /*form*/ typeof (element["onreset"]) is "function"
PASS /*form*/ typeof (element["onselect"]) is "function"
PASS /*form*/ typeof (element["onload"]) is "function"
PASS /*input*/ typeof (element["onclick"]) is "function"
PASS /*input*/ typeof (element["oncontextmenu"]) is "function"
PASS /*input*/ typeof (element["ondblclick"]) is "function"
PASS /*input*/ typeof (element["onmousedown"]) is "function"
PASS /*input*/ typeof (element["onmousemove"]) is "function"
PASS /*input*/ typeof (element["onmouseout"]) is "function"
PASS /*input*/ typeof (element["onmouseover"]) is "function"
PASS /*input*/ typeof (element["onmouseup"]) is "function"
PASS /*input*/ typeof (element["onmousewheel"]) is "function"
PASS /*input*/ typeof (element["onfocus"]) is "function"
PASS /*input*/ typeof (element["onblur"]) is "function"
PASS /*input*/ typeof (element["onkeydown"]) is "function"
PASS /*input*/ typeof (element["onkeypress"]) is "function"
PASS /*input*/ typeof (element["onkeyup"]) is "function"
PASS /*input*/ typeof (element["onscroll"]) is "function"
PASS /*input*/ typeof (element["onbeforecut"]) is "function"
PASS /*input*/ typeof (element["oncut"]) is "function"
PASS /*input*/ typeof (element["onbeforecopy"]) is "function"
PASS /*input*/ typeof (element["oncopy"]) is "function"
PASS /*input*/ typeof (element["onbeforepaste"]) is "function"
PASS /*input*/ typeof (element["onpaste"]) is "function"
PASS /*input*/ typeof (element["ondragenter"]) is "function"
PASS /*input*/ typeof (element["ondragover"]) is "function"
PASS /*input*/ typeof (element["ondragleave"]) is "function"
PASS /*input*/ typeof (element["ondrop"]) is "function"
PASS /*input*/ typeof (element["ondragstart"]) is "function"
PASS /*input*/ typeof (element["ondrag"]) is "function"
PASS /*input*/ typeof (element["ondragend"]) is "function"
PASS /*input*/ typeof (element["onselectstart"]) is "function"
PASS /*input*/ typeof (element["onsubmit"]) is "function"
PASS /*input*/ typeof (element["onerror"]) is "function"
PASS /*input*/ typeof (element["oninput"]) is "function"
PASS /*input*/ typeof (element["oninvalid"]) is "function"
PASS /*input*/ typeof (element["ontouchstart"]) is "function"
PASS /*input*/ typeof (element["ontouchmove"]) is "function"
PASS /*input*/ typeof (element["ontouchend"]) is "function"
PASS /*input*/ typeof (element["ontouchcancel"]) is "function"
PASS /*input*/ typeof (element["onwebkitfullscreenchange"]) is "function"
PASS /*input*/ typeof (element["onwebkitfullscreenerror"]) is "function"
PASS /*input*/ typeof (element["onabort"]) is "function"
PASS /*input*/ typeof (element["onchange"]) is "function"
PASS /*input*/ typeof (element["onreset"]) is "function"
PASS /*input*/ typeof (element["onselect"]) is "function"
PASS /*input*/ typeof (element["onload"]) is "function"
PASS successfullyParsed is true

TEST COMPLETE
